# backend/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "info": {
    "title": "DSNP Gateway Prototype API Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/profiles/{dsnpId}": {
      "get": {
        "operationId": "getProfile",
        "summary": "Get profile information for a specific user",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "createProfile",
        "summary": "Create/Edit the profile information for a current user",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditProfileRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "tokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "accessToken"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token invalid or not found"
      }
    },
    "schemas": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "fromId": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "content": {
            "description": "JSON-encoded Activity Content Note",
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the post"
          },
          "displayHandle": {
            "type": "string"
          }
        },
        "required": [
          "fromId",
          "contentHash",
          "content",
          "timestamp"
        ]
      },
      "EditProfileRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}